+++ src/client.ts
@@ -1,5 +1,5 @@
 import { randomUUID } from "crypto";
-import { CookieJar } from "./cookies";
+import { CookieJar, WebsiteData } from "./cookies";
 import { httpsGET } from "./http";
 import { Logger, NullLogger } from "./logger";
 import {
@@ -91,13 +91,37 @@ export class SymplifySDK {
             return null;
         }
 
+        const siteData = new WebsiteData(this.websiteID, cookies);
+        if (!siteData.isCompatible()) {
+            this.log.warn("findVariation: unsupported cookie generation");
+            return null;
+        }
+
         const project = findProjectWithName(this.config.latest, projectName);
         if (!project) {
             this.log.error(`findVariation: unknown project: ${projectName}`);
             return null;
         }
 
-        const visID = ensureVisitorID(cookies.get, cookies.set, this.websiteID, this.idGenerator);
+        if (project.state !== "active") {
+            return null;
+        }
+
+        const currAllocation = siteData.getAllocation(project);
+
+        switch (currAllocation) {
+            case undefined:
+                // no allocation data
+                break;
+            case null:
+                // explicit null allocation
+                return null;
+            default:
+                // variation allocation
+                return currAllocation.name;
+        }
+
+        const visID = ensureVisitorID(siteData, this.idGenerator);
         if (!visID) {
             this.log.error("could not get or generate a visitor ID");
             return null;
@@ -105,11 +129,15 @@ export class SymplifySDK {
 
         const variation = findVariationForVisitor(project, visID);
 
-        if (!variation) {
-            return null;
+        if (variation) {
+            siteData.rememberAllocation(project, variation);
+        } else {
+            siteData.rememberNullAllocation(project);
         }
 
-        return variation.name;
+        siteData.save(cookies);
+
+        return variation ? variation.name : null;
     }
 
     configURL(): string {